No longer make guesses when determining who provides a given feature
The new function `epkg-provided-by' returns the package that provides
a given feature.  It replaces `epkg--required', which did the same
thing, but also also guessed the package for certain features, those
involving autoload, version, and test libraries, in case the Epkg
database did not contain any information on the feature.  The new
function doesn't do that because this is no longer necessary now that
all those features are accounted for in the database.