epkg-provided: exclude bundled libraries by default
Unless the new optional argument INCLUDE-BUNDLED is non-nil, return
no bundled libraries.

This also affects the return value of `epkg-reverse-dependencies',
which now no longer includes invalid reverse dependencies that are
due to a package bundling libraries it shouldn't be bundling.

That in turn affects `epkg-insert-reverse-dependencies', which is
used when presenting a package description.  This is what motivated
this change, but it also makes sense elsewhere.