epkg-update: do not pull submodules also
It wasn't documented that we used to do so and it is potentially
highly undesirable to do it.  Updating all submodules can take a
long time, which also can be undesirable when you just want to
quickly update the database.  And finally if we continued to pull
submodules, then we would have to display progress information,
which is hard to do because we need to do it synchronously.

Re #4.