Write down the used table schemata
Add new library "epkg-schemata.el", defining `closql--db-init' for
`epkg-database' objects and the constant `epkg--db-table-schemata'.

`closql--db-init' used to be defined for `closql-database'.  That
method generated the database schemata from the relevant object class,
but was not flexible enough and was removed.  Our `closql--db-init'
method can be used to initialize a new Epkg database, but since that
already exists, all of this serves mostly as documentation.