Import find-file commands and various utilities from Emir
These commands and functions were previously defined in Emir, but they
are useful for other users beside the Emacsmirror maintainer.

In some cases the functions already used the "epkg-" prefix, in other
cases it is being changed from "emir-" to "epkg-".  Those are the only
changes; more adjustments follow in the next commit.