epkg-add-org-header: Document and fix
When executing the code blocks in a buffer, then the post-processing
block gets also executed in its own right.  This cannot be prevented
with ':eval no' because then it also does not get executed when it
is used for post-processing.  Luckily we can detect direct execution
because then ROWS is the empty string (the default value, set using
':var rows=\"\"').